test: unit tests for log getSimplified
